requirements.txt
openai
python-dotenv
fastapi
pydantic
pandas
matplotlib
numpy
tensorflow==2.16.1
pillow